% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{all_pos_ints}
\alias{all_pos_ints}
\title{Check whether all arguments are positive integers}
\usage{
all_pos_ints(x)
}
\arguments{
\item{x}{a vector containing the elements to be tested.}
}
\value{
Returns \code{TRUE} or \code{FALSE} accordingly.
}
\description{
\code{all_pos_ints} checks whether all the elements in a vector
  are positive integers.
}
\keyword{internal}
