% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{check_data}
\alias{check_data}
\title{Check the data is in the correct form}
\usage{
check_data(data, p)
}
\arguments{
\item{data}{a matrix or class \code{'ts'} object with \code{d>1} columns. Each column is taken to represent
a single time series. \code{NA} values are not supported.}

\item{p}{a positive integer specifying the autoregressive order of the model.}
}
\value{
Checks the data and tries to correct it. Throws an error if something is wrong and
  returns the corrected data otherwise.
}
\description{
\code{check_data} checks the data.
}
