% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{check_gmvar}
\alias{check_gmvar}
\title{Checks whether the given object has class attribute "gmvar" or "sgmvar"}
\usage{
check_gmvar(object)
}
\arguments{
\item{object}{S3 object to be tested}
}
\value{
Throws an error if the object doesn't have the class attribute "gmvar" or "sgmvar".
}
\description{
\code{check_gmvar} checks that the object has class attribute "gmvar" or "sgmvar".
}
