% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{check_gsmvar}
\alias{check_gsmvar}
\title{Checks whether the given object has class attribute 'gsmvar'}
\usage{
check_gsmvar(object, object_name)
}
\arguments{
\item{object}{S3 object to be tested}

\item{object_name}{what is the name of the object that should of class 'gsmvar'?}
}
\value{
Throws an error if the object doesn't have the class attribute 'gsmvar'.
}
\description{
\code{check_gsmvar} checks that the object has class attribute 'gsmvar'.
}
\keyword{internal}
