% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{check_pMd}
\alias{check_pMd}
\title{Check that p, M, and d are correctly set}
\usage{
check_pMd(p, M, d)
}
\arguments{
\item{p}{a positive integer specifying the autoregressive order of the model.}

\item{M}{a positive integer specifying the number of mixture components.}

\item{d}{the number of time series in the system.}
}
\value{
Throws an error if something is wrong.
}
\description{
\code{check_pMd} checks the arguments p, M, and d.
}
