% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcal.R
\name{create_J_matrix}
\alias{create_J_matrix}
\title{Create a special matrix J}
\usage{
create_J_matrix(d, p)
}
\arguments{
\item{d}{An integer representing the dimension of the identity matrix.}

\item{p}{An integer representing the factor by which to extend the matrix with zeros.}
}
\value{
A \eqn{d x dp} matrix \eqn{J} where the first \eqn{d x d} block is the identity matrix \eqn{I_d},
  and the rest is filled with zeros.
}
\description{
\code{create_J_matrix} generates a d x dp matrix J, where the first d x d block is the identity matrix I_d,
and the rest is filled with zeros.
}
\keyword{internal}
