% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{format_valuef}
\alias{format_valuef}
\title{Function factory for value formatting}
\usage{
format_valuef(digits)
}
\arguments{
\item{digits}{the number of decimals to print}
}
\value{
Returns a function that takes an atomic vector as argument
  and returns it formatted to character with \code{digits} decimals.
}
\description{
\code{format_valuef} is a function factory for
  formatting values with certain number of digits.
}
\keyword{internal}
