% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdpdef}
\alias{gdpdef}
\title{U.S. real GDP percent change and GDP implicit price deflator percent change.}
\format{
A numeric matrix of class \code{'ts'} with 244 rows and 2 columns with one time series in each column:
\describe{
  \item{First column (GDP):}{The quarterly percent change of real U.S. GDP, from 1959Q1 to 2019Q4, \url{https://fred.stlouisfed.org/series/GDPC1}.}
  \item{Second column (GDPDEF):}{The quarterly percent change of U.S. GDP implicit price deflator, from 1959Q1 to 2019Q4, \url{https://fred.stlouisfed.org/series/GDPDEF}.}
}
}
\source{
The Federal Reserve Bank of St. Louis database
}
\usage{
gdpdef
}
\description{
A dataset containing a quarterly U.S. time series with two components:
the percentage change of real GDP and the percentage change of GDP implicit price deflator,
covering the period from 1959Q1 - 2019Q4.
}
\keyword{datasets}
