% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{get_IC}
\alias{get_IC}
\title{Calculate AIC, HQIC, and BIC}
\usage{
get_IC(loglik, npars, obs)
}
\arguments{
\item{loglik}{log-likelihood value}

\item{npars}{number of (freely estimated) parameters in the model}

\item{obs}{numbers of observations with starting values excluded for conditional models.}
}
\value{
Returns a data frame containing the information criteria values.
}
\description{
\code{get_IC} calculates the information criteria values AIC, HQIC, and BIC.
}
\details{
Note that for conditional models with different autoregressive order p the
 information criteria values are \strong{NOT} comparable.
}
