% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcal.R
\name{get_symmetric_sqrt}
\alias{get_symmetric_sqrt}
\title{Calculate symmetric square root matrix of a positive definite covariance
  matrix}
\usage{
get_symmetric_sqrt(Omega)
}
\arguments{
\item{Omega}{a positive definite covariance matrix}
}
\value{
a vectorized symmetric square root matrix of the matrix \code{Omega}.
}
\description{
\code{get_symmetric_sqrt} calculates symmetric square root matrix
 of a positive definite covariance matrix
}
\section{Warning}{

 No argument checks!
}

\keyword{internal}
