% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwardCompatibility.R
\name{gmvar_to_gsmvar}
\alias{gmvar_to_gsmvar}
\title{Makes class 'gmvar' objects compatible with the functions using class 'gsmvar' objects}
\usage{
gmvar_to_gsmvar(gsmvar)
}
\arguments{
\item{gsmvar}{a class 'gmvar' or 'gsmvar' object.}
}
\value{
If the provided object has the class 'gsmvar', the provided object
  is returned without modifications. If the provided object has the class 'gmvar',
  its element \code{$model} is given a new subelement called also model and this is
  set to be "GMVAR". Also, the class of this object is changes to 'gsmvar' and then
  it is returned.
}
\description{
\code{gmvar_to_gsmvar} class 'gmvar' objects compatible with the functions using
s class 'gsmvar' objects
}
\details{
This exists so that models estimated with earlier versions
  of the package can be used conveniently.
}
