% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmvarkit.R
\docType{package}
\name{gmvarkit-package}
\alias{gmvarkit}
\alias{gmvarkit-package}
\title{gmvarkit: Estimate Gaussian and Student's t Mixture Vector Autoregressive Models}
\description{
\code{gmvarkit} is a package for reduced form and structural Gaussian mixture vector
  autoregressive (GMVAR), Student's t Mixture Vector Autoregressive (StMVAR),
  or Gaussian and Student's t Mixture Vector Autoregressive (G-StMVAR) model analysis.
  It provides functions for unconstrained and constrained maximum likelihood estimation of the model parameters,
  quantile residuals tests, graphical diagnostics, estimation of generalized impulse response function,
  estimation of generalized forecast error variance decomposition, simulation from GMVAR processes,
  forecasting, and more.

  The readme file is a good place to start and the vignette might be useful too.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/saviviro/gmvarkit/issues}
}

}
\author{
\strong{Maintainer}: Savi Virolainen \email{savi.virolainen@helsinki.fi} (\href{https://orcid.org/0000-0002-5075-6821}{ORCID})

}
