% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmvarkit.R
\docType{package}
\name{gmvarkit}
\alias{gmvarkit}
\alias{gmvarkit-package}
\title{gmvarkit: Estimate Gaussian Mixture Vector Autoregressive (GMVAR) model}
\description{
\code{gmvarkit} is a package for estimating Gaussian Mixture Vector Autoregressive (GMVAR) model-
  It provides functions for quantile residuals tests, graphical diagnostics, forecasting and simulations. Applying
  general linear constraints on the autoregressive parameters is supported.

  Most of the functions documented are not exported, but for internal use only. The vignette
  is a good place to start.
}
