% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearIRF.R
\name{linear_IRF}
\alias{linear_IRF}
\title{Estimate linear impulse response function based on a single regime of a structural GMVAR,
  StMVAR, or G-StMVAR model.}
\usage{
linear_IRF(
  gsmvar,
  N = 30,
  regime = 1,
  which_cumulative = numeric(0),
  scale = NULL,
  ci = NULL,
  bootstrap_reps = NULL,
  ncores = 2,
  seeds = NULL,
  ...
)
}
\arguments{
\item{gsmvar}{an object of class \code{'gsmvar'} defining a structural or reduced form
GSMVAR model. For a reduced form model, the shocks are automatically identified by
the lower triangular Cholesky decomposition.}

\item{N}{a positive integer specifying the horizon how far ahead should the
linear impulse responses be calculated.}

\item{regime}{Based on which regime the linear IRF should be calculated?
An integer in \eqn{1,...,M}.}

\item{which_cumulative}{a numeric vector with values in \eqn{1,...,d}
(\code{d=ncol(data)}) specifying which the variables for which the linear impulse
responses should be cumulative. Default is none.}

\item{scale}{should the linear IRFs to some of the shocks be scaled so that they
correspond to a specific magnitude of instantaneous response of some specific
variable? Provide a length three vector where the shock of interest
is given in the first element (an integer in \eqn{1,...,d}), the variable of
interest is given in the second element (an integer in \eqn{1,...,d}), and
the magnitude of its instantaneous response in the third element (a non-zero real number).
If the linear IRFs of multiple shocks should be scaled, provide a matrix which has one
column for each of the shocks with the columns being the length three vectors described above.}

\item{ci}{a real number in \eqn{(0, 1)} specifying the confidence level of the
confidence intervals calculated via a fixed-design wild residual bootstrap method.
Available only for models that impose linear autoregressive dynamics
(excluding changes in the volatility regime).}

\item{bootstrap_reps}{the number of bootstrap repetitions for estimating confidence bounds.}

\item{ncores}{the number of CPU cores to be used in parallel computing when bootstrapping confidence bounds}

\item{seeds}{a numeric vector of length \code{bootstrap_reps} initializing the seed for the random
generator for each bootstrap replication.}

\item{...}{parameters passed to the plot method \code{plot.irf} that plots
the results.}
}
\value{
Returns a class \code{'irf'} list with the linear IRFs in ... FILL IN!
}
\description{
\code{linear_IRF} estimates linear impulse response function based on a single regime
  of a structural GMVAR, StMVAR, or G-StMVAR model.
}
\details{
The model DOES NOT need to be structural in order for this function to be
  applicable. When an identified structural GMVAR, StMVAR, or G-StMVAR model is
  provided in the argument \code{gsmvar}, the identification imposed by the model
  is used. When a reduced form model is provided in the argument \code{gsmvar},
  lower triangular Cholesky identification is used to identify the shocks.

  FILL IN DETAILS ABOUT CONFIDENCE INTERVALS
}
\examples{
 # FILL IN
}
\references{
\itemize{
   \item Kilian L. and Lütkepohl H. 2017. Structural Vectors Autoregressive Analysis.
         \emph{Cambridge University Press}, Cambridge.
 }
}
\seealso{
\code{\link{GIRF}}, \code{\link{GFEVD}}, \code{\link{fitGSMVAR}}, \code{\link{GSMVAR}},
  \code{\link{gsmvar_to_sgsmvar}}, \code{\link{reorder_W_columns}}, \code{\link{swap_W_signs}}
}
