% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcal.R
\name{mat_power}
\alias{mat_power}
\title{Compute the j:th power of a square matrix A}
\usage{
mat_power(A, j)
}
\arguments{
\item{A}{A square numeric matrix.}

\item{j}{A natural number representing the power to which the matrix will be raised.}
}
\value{
A matrix which is A raised to the power j.
}
\description{
\code{mat_power} computes the j:th power of a square matrix A using
exponentiation by squaring.
}
\keyword{internal}
