% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwardCompatibility.R
\name{plot.gmvarpred}
\alias{plot.gmvarpred}
\alias{print.gmvarpred}
\title{plot method for class 'gmvarpred' objects}
\usage{
\method{plot}{gmvarpred}(x, ..., nt, mix_weights = TRUE, add_grid = TRUE)

\method{print}{gmvarpred}(x, ..., digits = 2)
}
\arguments{
\item{x}{object of class \code{'gsmvarpred'} generated by \code{predict.gsmvar}.}

\item{...}{arguments passed to \code{grid} which plots grid to the figure.}

\item{nt}{a positive integer specifying the number of observations to be plotted
along with the prediction (ignored if \code{plot_res==FALSE}). Default is \code{round(nrow(data)*0.15)}.}

\item{mix_weights}{\code{TRUE} if forecasts for mixing weights should be plotted,
\code{FALSE} in not.}

\item{add_grid}{should grid be added to the plots?}

\item{digits}{how many digits to print?}
}
\description{
\code{plot.gmvarpred} is plot method for gsmvarpred objects.
 EXISTS FOR BACKWARD COMPATIBILITY. THE CLASS 'gmvarpred' IS DEPRECATED FROM
 THE VERSION 2.0.0 ONWARD: WE USE THE CLASS 'gsmvarpred' NOW.
}
\details{
These methods exist so that objects created with earlier versions
  of the package can be used normally.
}
