% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecatedFunctions.R
\name{print.gmvar}
\alias{print.gmvar}
\alias{summary.gmvar}
\title{Deprecated S3 methods for the deprecated class 'gmvar'}
\usage{
\method{print}{gmvar}(gmvar, ..., digits = 2)

\method{summary}{gmvar}(gmvar)

\method{summary}{gmvar}(gmvar)
}
\arguments{
\item{gmvar}{a class 'gmvar' object. THIS CLASS IS DEPRECATED FROM THE VERSION
2.0.0 ONWARDS.}

\item{digits}{number of digits to be printed.}
}
\description{
Deprecated S3 methods for the deprecated class 'gmvar'. From
  the gmvarkit version 2.0.0 onwards, class 'gsmvar' is used instead.
}
\details{
These methods exists so that models estimated with earlier versions
  of the package can be used.
}
