% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.gmvarpred}
\alias{print.gmvarpred}
\title{Print method for class 'gmvarpred' objects}
\usage{
\method{print}{gmvarpred}(x, ..., digits = 2)
}
\arguments{
\item{x}{object of class \code{'gmvarpred'} generated by \code{predict.gmvar}.}

\item{...}{currently not used.}

\item{digits}{the number of decimals to print}
}
\description{
\code{print.gmvarpred} is a print method for object generated
 by \code{predict.gmvar}.
}
\examples{
# GMVAR(2, 2), d=2 model;
params22 <- c(0.36, 0.121, 0.223, 0.059, -0.151, 0.395, 0.406, -0.005,
 0.083, 0.299, 0.215, 0.002, 0.03, 0.484, 0.072, 0.218, 0.02, -0.119,
 0.722, 0.093, 0.032, 0.044, 0.191, 1.101, -0.004, 0.105, 0.58)
mod22 <- GMVAR(gdpdef, p=2, M=2, params=params22)
pred22 <- predict(mod22, n_ahead=3, plot_res=FALSE)
print(pred22)
print(pred22, digits=3)
}
