% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwardCompatibility.R
\name{print.gmvarsum}
\alias{print.gmvarsum}
\title{Summary print method from objects of class 'gmvarsum'}
\usage{
\method{print}{gmvarsum}(x, ..., digits)
}
\arguments{
\item{x}{object of class 'gsmvarsum' generated by \code{summary.gsmvar}.}

\item{...}{currently not used.}

\item{digits}{the number of digits to be printed.}
}
\description{
\code{print.gmvarsum} is a print method for object \code{'gmvarsum'}.
  EXISTS FOR BACKWARD COMPATIBILITY. CLASS 'gmvarsum' IS DEPRECATED FROM THE VERSION
  2.0.0. ONWARDS. NOW, WE USE THE CLASS 'gsmvarsum'.
}
