% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.hypotest}
\alias{print.hypotest}
\title{Print method for the class hypotest}
\usage{
\method{print}{hypotest}(x, ..., digits = 4)
}
\arguments{
\item{x}{object of class \code{'hypotest'} generated by the function \code{Wald_test} or \code{LR_test}.}

\item{...}{currently not in use.}

\item{digits}{how many significant digits to print?}
}
\description{
\code{print.hypotest} is the print method for the class hypotest
 objects.
}
