% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcal.R
\name{re_decompose_Omegas}
\alias{re_decompose_Omegas}
\title{In the decomposition of the covariance matrices (Muirhead, 1982, Theorem A9.9), change
  covariance matrix that is the first one, or in more general put them to new ordeer.}
\usage{
re_decompose_Omegas(d, M, W, lambdas, new_perm = 1:M)
}
\arguments{
\item{d}{the number of time series in the system.}

\item{M}{a positive integer specifying the number of mixture components.}

\item{W}{a length \code{d^2} vector containing the vectorized W matrix.}

\item{lambdas}{a length \code{d*(M-1)} vector of the form \strong{\eqn{\lambda_{2}}}\eqn{,...,}\strong{\eqn{\lambda_{M}}}
where \strong{\eqn{\lambda_{m}}}\eqn{=(\lambda_{m1},...,\lambda_{md})}}

\item{new_perm}{a vector of length \code{M} giving the new order of the covarince matrices
(relative to the current order)}
}
\value{
Returns a \eqn{d^2 + (M - 1)*d x 1} vector of the form \code{c(vec(new_W), new_lambdas)}
  where the lambdas parameters are in the regimewise order (first regime 2, then 3, etc) and the
  "new W" and "new lambdas" are constitute the new decomposition with the order of the covariance
  matrices given by the argument \code{new_perm}. Notice that if the first element of \code{new_perm}
  is one, the W matrix will be the same and the lambdas are just re-ordered.
}
\description{
\code{re_decompose_Omegas} exchanges the order of the covariance matrices in
  the decomposition of Muirhead (1982, Theorem A9.9) and returns the new decomposition.
}
\details{
We consider the following decomposition of positive definite covariannce matrices:
 \eqn{\Omega_1 = WW'}, \eqn{\Omega_m = W\Lambda_{m}W'}, \eqn{m=2,..,M} where \eqn{\Lambda_{m} = diag(\lambda_{m1},...,\lambda_{md})}
 contains the strictly postive eigenvalues of \eqn{\Omega_m\Omega_1^{-1}} and the column of the invertible \eqn{W} are the
 corresponding eigenvectors. Note that this decomposition does not necessarily exists for \eqn{M > 2}.

 See Muirhead (1982), Theorem A9.9 for more details on the decomposition and the source code for more details on the reparametrization.
}
\section{Warning}{

 No argument checks! Does not work with dimension \eqn{d=1} or with only
 one mixture component \eqn{M=1}.
}

\examples{

}
\references{
\itemize{
  \item Muirhead R.J. 1982. Aspects of Multivariate Statistical Theory, \emph{Wiley}.
}
}
