% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcal.R
\name{unvec}
\alias{unvec}
\title{Reverse vectorization operator}
\usage{
unvec(d, a)
}
\arguments{
\item{d}{the number of rows in the square matrix to be formed.}

\item{a}{a size \eqn{(d^2x1)} vector to be unvectorized into a \eqn{(dxd)} matrix.}
}
\value{
a matrix of size \eqn{(dxd)}.
}
\description{
\code{unvec} forms a square matrix from a vector of
 stacked columns, stacked by \code{vec}.
}
\section{Warning}{

 No argument checks!
}

\keyword{internal}
