% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcal.R
\name{unvech}
\alias{unvech}
\title{Reverse operator of the parsimonious vectorization operator \code{vech}}
\usage{
unvech(d, a)
}
\arguments{
\item{d}{number of rows the square matrix to be formed.}

\item{a}{a size \eqn{(d(d+1)/2x1)} vector to be unvectorized into a symmetric \eqn{(dxd)} matrix.}
}
\value{
a symmetric matrix of size \eqn{(dxd)}.
}
\description{
\code{unvech} creates a symmetric matrix from the given vector by
  copying the lower triangular part to be the upper triangular part as well.
}
\section{Warning}{

 No argument checks!
}

\keyword{internal}
