% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncondMoments.R
\name{VAR_init_acov}
\alias{VAR_init_acov}
\title{Calculate the initial (lag zero)  dp-dimensional autocovariance matrix of p consecutive
 observations of VAR process}
\usage{
VAR_init_acov(p, d, all_Am, Omega_m)
}
\arguments{
\item{p}{a positive integer specifying the autoregressive order of the model.}

\item{all_Am}{HERE PARAMETERS}

\item{Omega_m}{the dxd error term covarian}
}
\description{
\code{VAR_init_acov} calculate the initial (lag zero) dp-dimensional autocovariance
 matrix of a VAR process (that is, for p consecutive observations) by taking advantage of the algorithm
 proposed by McElroy (2017).
}
\details{
Copyright (2015) Tucker McElroy

This program is free software; you can redistribute it and/or
modify it under the terms of the GNU General Public License
as published by the Free Software Foundation; either version 2
of the License, or (at your option) any later version.

This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

You should have received a copy of the GNU General Public License
along with this program; if not, write to the Free Software
Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
}
\references{
HERE THE REFERENCE
}
