% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcal.R
\name{vec}
\alias{vec}
\title{Vectorization operator}
\usage{
vec(A)
}
\arguments{
\item{A}{a size \eqn{(dxd)} square matrix to be vectorized.}
}
\value{
a vector of size \eqn{(d^2x1)}.
}
\description{
\code{vec} stacks columns of the given matrix to form a vector.
}
\section{Warning}{

 No argument checks!
}

\keyword{internal}
