% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcal.R
\name{vech}
\alias{vech}
\title{Parsimonious vectorization operator for symmetric matrices}
\usage{
vech(A)
}
\arguments{
\item{A}{a size \eqn{(dxd)} symmetric matrix to be vectorized parsimoniously.}
}
\value{
a vector of size \eqn{(d(d+1)/2x1)}.
}
\description{
\code{vech} stacks columns of the given matrix from
  the principal diagonal downwards (including elements on the diagonal) to form a vector.
}
\section{Warning}{

 No argument checks!
}

\keyword{internal}
