% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcal.R
\name{Wvec}
\alias{Wvec}
\title{Vectorization operator that removes zeros}
\usage{
Wvec(W)
}
\arguments{
\item{W}{a size \eqn{(dxd)} square matrix to be vectorized.}
}
\value{
a vector of length \eqn{d^2 - n_zeros} where \eqn{n_zeros} is the
  number of zero entries in the matrix \code{W}.
}
\description{
\code{Wvec} stacks columns of the given matrix to form a vector
  and removes elements that are zeros.
}
\section{Warning}{

 No argument checks!
}

\keyword{internal}
