% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{check_null_data}
\alias{check_null_data}
\title{Checks whether the given object contains data}
\usage{
check_null_data(gsmvar)
}
\arguments{
\item{gsmvar}{an object of class \code{'gsmvar'}, typically created with \code{fitGSMVAR} or \code{GSMVAR}.}
}
\value{
Throws an error if is.null(gsmvar$data).
}
\description{
\code{check_null_data} checks that the gsmvar object has data.
}
\keyword{internal}
