% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morePlots.R
\name{cond_moment_plot}
\alias{cond_moment_plot}
\title{Conditional mean or variance plot for a GMVAR model}
\usage{
cond_moment_plot(
  gmvar,
  which_moment = c("mean", "variance"),
  grid = FALSE,
  ...
)
}
\arguments{
\item{gmvar}{an object of class \code{'gmvar'} created with \code{fitGMVAR} or \code{GMVAR}.}

\item{which_moment}{should conditional means or variances be plotted?}

\item{grid}{add grid to the plots?}

\item{...}{additional paramters passed to \code{grid(...)} plotting the grid if \code{grid == TRUE}.}
}
\description{
\code{cond_moment_plot} plots the one-step in-sample conditional means/variances of the model along with
 the individual time series contained in the model (e.g. the time series the model was fitted to). Also plots
 the regimewise conditional means/variances multiplied with mixing weights.
}
\details{
The conditional mean plot works best if the data contains positive values only.
 \code{acf} from the package \code{stats} and the plot method for class \code{'acf'} objects is employed.
}
\examples{
# These examples use the data 'eurusd' which comes with the
# package, but in a scaled form.
data <- cbind(10*eurusd[,1], 100*eurusd[,2])
colnames(data) <- colnames(eurusd)

# GMVAR(2,2), d=2 model:
params222 <-  c(1.386, -0.765, 1.314, 0.145, 0.094, 1.292, -0.389,
 -0.070, -0.109, -0.281, 0.920, -0.025, 4.839, 1.005, 5.928, 1.248,
  0.077, -0.040, 1.266, -0.272, -0.074, 0.034, -0.313, 5.855, 3.570,
  9.838, 0.740)
mod222 <- GMVAR(data, p=2, M=2, params=params222)

cond_moment_plot(mod222, which_moment="mean")
cond_moment_plot(mod222, which_moment="variance")
cond_moment_plot(mod222, which_moment="mean", grid=TRUE, lty=2)
}
\references{
\itemize{
   \item Kalliovirta L., Meitz M. and Saikkonen P. 2016. Gaussian mixture vector autoregression.
         \emph{Journal of Econometrics}, \strong{192}, 485-498.
   \item Virolainen S. 2020. Structural Gaussian mixture vector autoregressive model. Unpublished working
     paper, available as arXiv:2007.04713.
 }
}
\seealso{
\code{\link{profile_logliks}}, \code{\link{fitGMVAR}}, \code{\link{GMVAR}}, \code{\link{quantile_residual_tests}},
 \code{\link{LR_test}}, \code{\link{Wald_test}}, \code{\link{diagnostic_plot}}
}
