% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmvarkit.R
\name{gmvarkit}
\alias{gmvarkit}
\title{gmvarkit: Estimate Gaussian Mixture Vector Autoregressive (GMVAR) model}
\description{
\code{gmvarkit} is a package for reduced form and structural Gaussian mixture vector
  autoregressive (GMVAR) model analysis.
  It provides functions for unconstrained and constrained maximum likelihood estimation of the model parameters,
  quantile residuals tests, graphical diagnostics, estimation of generalized impulse response function,
  simulation from GMVAR processes, forecasting, and more.

  Many of the functions documented are not exported but intended internal use only. The readme file
  is a good place to start and the vignette might be useful too.
}
