% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearIRF.R
\name{linear_IRF}
\alias{linear_IRF}
\title{Estimate linear impulse response function based on a single regime of a structural GMVAR,
  StMVAR, or G-StMVAR model.}
\usage{
linear_IRF(
  gsmvar,
  which_shocks,
  N = 30,
  regime = 1,
  which_cumulative = numeric(0),
  scale = NULL,
  scale_type = c("instant", "peak"),
  scale_horizon = N,
  ci = c(0.95, 0.8),
  seed,
  ...
)
}
\arguments{
\item{gsmvar}{an object of class \code{'gsmvar'} defining a structural or reduced form
GSMVAR model. For a reduced form model, the shocks are automatically identified by
the lower triangular Cholesky decomposition.}

\item{which_shocks}{a numeric vector of length at most \eqn{d}
(\code{=ncol(data)}) and elements in \eqn{1,...,d} specifying the
structural shocks for which the IRF should be estimated.}

\item{N}{a positive integer specifying the horizon how far ahead should the
linear impulse responses be calculated.}

\item{regime}{Based on which regime the linear IRF should be calculated?
An integer in \eqn{1,...,M}.}

\item{which_cumulative}{a numeric vector with values in \eqn{1,...,d}
(\code{d=ncol(data)}) specifying which the variables for which the linear impulse
responses should be cumulative. Default is none.}

\item{scale}{should the linear IRFs to some of the shocks be scaled so that they
correspond to a specific magnitude of instantaneous or peak response
of some specific variable (see the argument \code{scale_type})?
Provide a length three vector where the shock of interest
is given in the first element (an integer in \eqn{1,...,d}), the variable of
interest is given in the second element (an integer in \eqn{1,...,d}), and
the magnitude of its instantaneous or peak response in the third element
(a non-zero real number). If the linear IRFs of multiple shocks should be scaled,
provide a matrix which has one column for each of the shocks with the columns being
the length three vectors described above.}

\item{scale_type}{If argument \code{scale} is specified, should the linear IRFs be
scaled to match an instantaneous response (\code{"instant"}) or peak response
(\code{"peak"}). If \code{"peak"}, the scale is based on the largest magnitude
of peak response in absolute value. Ignored if \code{scale} is not specified.}

\item{scale_horizon}{If \code{scale_type == "peak"} what the maximum horizon up
to which peak response is expected? Scaling won't based on values after this.}

\item{ci}{a numeric vector with elements in \eqn{(0, 1)} specifying the
confidence levels of the confidence intervals calculated via a bootstrap
method, see the details section.}

\item{seed}{a length one numeric vector initializing the seed for the random generator.}

\item{...}{parameters passed to the plot method \code{plot.irf} that plots
the results.}
}
\value{
Returns a class \code{'irf'} list with the linear IRFs in ... FILL IN!
}
\description{
\code{linear_IRF} estimates linear impulse response function based on a single regime
  of a structural GMVAR, StMVAR, or G-StMVAR model.
}
\details{
The model DOES NOT need to be structural in order for this function to be
  applicable. When an identified structural GMVAR, StMVAR, or G-StMVAR model is
  provided in the argument \code{gsmvar}, the identification imposed by the model
  is used. When a reduced form model is provided in the argument \code{gsmvar},
  lower triangular Cholesky identification is used to identify the shocks.

  FILL IN DETAILS ABOUT CONFIDENCE INTERVALS
}
\examples{
 # FILL IN
}
\references{
\itemize{
   \item Kilian L. and Lütkepohl H. 2017. Structural Vectors Autoregressive Analysis.
         \emph{Cambridge University Press}, Cambridge.
 }
}
\seealso{
\code{\link{GIRF}}, \code{\link{GFEVD}}, \code{\link{fitGSMVAR}}, \code{\link{GSMVAR}},
  \code{\link{gsmvar_to_sgsmvar}}, \code{\link{reorder_W_columns}}, \code{\link{swap_W_signs}}
}
