% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\name{plot.gmvarpred}
\alias{plot.gmvarpred}
\title{plot method for class 'gmvarpred' objects}
\usage{
\method{plot}{gmvarpred}(x, ..., nt, add_grid = TRUE)
}
\arguments{
\item{x}{object of class \code{'gmvarpred'} generated by \code{predict.gmvar}.}

\item{...}{arguments passed to \code{grid} which plots grid to the figure.}

\item{nt}{a positive integer specifying the number of observations to be plotted
along with the prediction. Default is \code{round(nrow(data)*0.2)}.}

\item{add_grid}{should grid be added to the plots?}
}
\description{
\code{plot.gmvarpred} is plot method for gmvarpred objects.
}
\details{
This method is used plot forecasts of GMVAR processes
}
\references{
\itemize{
   \item Kalliovirta L., Meitz M. and Saikkonen P. 2016. Gaussian mixture vector autoregression.
         \emph{Journal of Econometrics}, \strong{192}, 485-498.
 }
}
