% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{warn_eigens}
\alias{warn_eigens}
\title{Warn about near-unit-roots in some regimes}
\usage{
warn_eigens(gmvar, tol = 0.002)
}
\arguments{
\item{gmvar}{an object of class \code{'gmvar'} created with \code{fitGMVAR} or \code{GMVAR}.}

\item{tol}{if eigenvalue is closer than \code{tol} to its bound, a warning is thrown}
}
\value{
Doesn't return anything.
}
\description{
\code{warn_ar_roots} warns if the model contains near-unit-roots in some regimes
}
\details{
Warns if, for some regime, some moduli of "bold A" eigenvalues are larger than \code{1 - tol} or
 some eigenvalue of the error term covariance matrix is smaller than \code{tol}.
}
